+++ src/helpers/rules.js
@@ -1,5 +1,7 @@
 const micromatch = require("micromatch");
 
+const { isArray, replaceObjectValuesInTemplates } = require("./utils");
+
 const REPO_URL = "https://github.com/javierbrea/eslint-plugin-boundaries";
 
 function removePluginNamespace(ruleName) {
@@ -43,9 +45,7 @@ function dependencyLocation(node, context) {
 }
 
 function micromatchPatternReplacingObjectValues(pattern, object) {
-  return Object.keys(object).reduce((result, objectKey) => {
-    return result.replace(`\${${objectKey}}`, object[objectKey]);
-  }, pattern);
+  return replaceObjectValuesInTemplates(pattern, object);
 }
 
 function isObjectMatch(objectWithMatchers, object, objectWithValuesToReplace) {
@@ -69,10 +69,10 @@ function rulesMainKey(key) {
 
 function ruleMatch(ruleMatchers, elementInfo, isMatch, elementToCompare = {}) {
   let match = { result: false, report: null };
-  const matchers = !Array.isArray(ruleMatchers) ? [ruleMatchers] : ruleMatchers;
+  const matchers = !isArray(ruleMatchers) ? [ruleMatchers] : ruleMatchers;
   matchers.forEach((matcher) => {
     if (!match.result) {
-      if (Array.isArray(matcher)) {
+      if (isArray(matcher)) {
         const [value, captures] = matcher;
         match = isMatch(elementInfo, value, captures, elementToCompare.capturedValues);
       } else {
@@ -110,9 +110,16 @@ function getElementRules(elementInfo, options, mainKey) {
     return [];
   }
   const key = rulesMainKey(mainKey);
-  return options.rules.filter((rule) => {
-    return ruleMatch(rule[key], elementInfo, isMatchElementType).result;
-  });
+  return options.rules
+    .map((rule, index) => {
+      return {
+        ...rule,
+        index,
+      };
+    })
+    .filter((rule) => {
+      return ruleMatch(rule[key], elementInfo, isMatchElementType).result;
+    });
 }
 
 function elementRulesAllowDependency({
@@ -122,12 +129,21 @@ function elementRulesAllowDependency({
   isMatch,
   rulesMainKey: mainKey,
 }) {
-  const [result, report] = getElementRules(element, options, mainKey).reduce(
+  const [result, report, ruleReport] = getElementRules(element, options, mainKey).reduce(
     (allowed, rule) => {
       if (rule.disallow) {
         const match = ruleMatch(rule.disallow, dependency, isMatch, element);
         if (match.result) {
-          return [false, match.report];
+          return [
+            false,
+            match.report,
+            {
+              element: rule[rulesMainKey(mainKey)],
+              disallow: rule.disallow,
+              index: rule.index,
+              message: rule.message || options.message,
+            },
+          ];
         }
       }
       if (rule.allow) {
@@ -138,11 +154,19 @@ function elementRulesAllowDependency({
       }
       return allowed;
     },
-    [options.default === "allow", null]
+    [
+      options.default === "allow",
+      null,
+      {
+        isDefault: true,
+        message: options.message,
+      },
+    ]
   );
   return {
     result,
     report,
+    ruleReport,
   };
 }
 
@@ -155,4 +179,5 @@ module.exports = {
   elementRulesAllowDependency,
   getElementRules,
   rulesMainKey,
+  micromatchPatternReplacingObjectValues,
 };