+++ CHANGELOG.md
@@ -11,6 +11,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 ### BREAKING CHANGES
 
+## [unreleased]
+
+### Added
+- feat(#87): Support custom messages in `element-types` config
+
+### Changed
+- feat(#87): Add context about the specific rule forbidding the import to the `element-types` error message. Add information about file and dependency when import is forbidden due to the default configuration.
+- refactor: Add isArray and isString utils
+- chore(deps): Support any eslint version greater than 6.x in peerDependencies
+
+### Fixed
+- fix: Support array of micromatch patterns when replacing captured values
+- docs: Fix broken links
+
+### Removed
+- feat: Remove cache traces from debug mode
+
 ## [2.5.1] - 2021-11-06
 
 ### Fixed